requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.5.0
opencv-python>=4.7.0
tqdm>=4.65.0
albumentations>=1.3.0
scikit-learn>=1.2.0
tensorboard>=2.13.0